+++ qwt/tests/cpudemo.py
@@ -285,8 +285,8 @@ class CpuCurve(QwtPlotCurve):
 class CpuPlot(QwtPlot):
     HISTORY = 60
 
-    def __init__(self, *args, unattended=False):
-        QwtPlot.__init__(self, *args)
+    def __init__(self, unattended=False):
+        QwtPlot.__init__(self)
 
         self.curves = {}
         self.data = {}